Removed non-wizdom operations from wizdom client
